+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # CHANGELOG for husky
 
+* dev-develop
+    * FEATURE     #737 Added possibility to define overlay spacing for each slide
+    * BUGFIX      #738 Removed action icon in datagrid if no icon is defined
+    * ENHANCEMENT #735 Added maxHeight options to toolbar dropdownOptions
+    * BUGFIX      #722 Fixed create hover button for ghost pages and made it optional
+    * FEATURE     #733 Created addImage functionality in dropzone
+    * ENHANCEMENT #723 Added description class to itembox and made content take up the whole width
+    * FEATURE     #714 Tabs and language Changer are now in a new section -> language Changer can be used without Tabs.
+    * BUGFIX      #729 Update parent checkbox style when removing child in table-view datagrid-decorator
+    * BUGFIX      #729 Remove children rows on parent deletion in table-view datagrid-decorator
+    * ENHANCEMENT #731 Added expandIds-option to the datagrid component
+    * ENHANCEMENT #721 Improved cropping in tiles view and fixed dropzone issue when destroying
+    * ENHANCEMENT #715 Display effective limit in dropdown-pagination decorator
+    * ENHANCEMENT #715 Replace expand-icon with loader when loading children in datagrid
+    * ENHANCEMENT #715 Append currently selected items to request when loading data in datagrid
+    * BUGFIX      #707 Added translations to titles in matrix
+    * BUGFIX      #718 Made infinite scroll pagination add all records at once if possible
+    * ENHANCEMENT #716 Made the reset to the "show all" only happen, when the datagrid was in the "show selected" state before
+    * ENHANCEMENT #713 Removed overlay from dropzone and implemented own overlay-style for dropzone
+    * ENHANCEMENT #712 Removed data-navigation from navigation component
+    * FEATURE     #706 Created tiles view for datagrid
+    * FEATURE     #705 Added placeholder option to ckeditor component
+
 * 0.25.1 (2016-09-15)
     * HOTFIX      #726 Added singleMarkable option and get-marked event
     * HOTFIX      #725 Added locale to pagination request of data-navigation